+++ resources/functions/init/Init-AutoUpdater.js
@@ -1,0 +1,15 @@
+const {app} = require('electron')
+const {autoUpdater} = require("electron-updater");
+autoUpdater.logger = require("electron-log");
+
+exports.InitializeAutoUpdater = function () {
+
+    if (app.config.advanced.autoUpdaterBetaBuilds) {
+        autoUpdater.allowPrerelease = true
+        autoUpdater.allowDowngrade = false
+    }
+    console.log("[AutoUpdater] Checking for updates...")
+    autoUpdater.checkForUpdatesAndNotify().then(r => console.log(`[AutoUpdater] Latest Version is ${r.updateInfo.version}`))
+
+
+}
