+++ resources/functions/init/Init-Logging.js
@@ -1,0 +1,12 @@
+const log = require("electron-log");
+
+exports.InitializeLogging = function () {
+
+
+    console.log('---------------------------------------------------------------------')
+    console.log('Apple-Music-Electron application has started.');
+    console.log("---------------------------------------------------------------------")
+    console.log = log.log;
+
+
+}
